requirements.txt
flask
tensorflow-cpu  <-- Gunakan versi CPU agar lebih ringan di hosting gratis
pillow
numpy
gunicorn